+++ packages/cli/api-importers/asyncapi-to-ir/src/converters/AbstractServerConverter.ts
@@ -47,6 +47,7 @@ export abstract class AbstractServerConverter<TServer> extends AbstractConverter
 
         return {
             defaultEnvironment: defaultEnvironmentId,
+            baseUrlEnvVar: undefined,
             environments: Environments.singleBaseUrl({
                 environments
             })